Added bloc pattern to the project

+++ lib/models/payment_model.dart
@@ -4,7 +4,7 @@ class PaymentModel {
   IconData _icon;
   String _name, _date, _hour;
   Color _color;
-  double _amount;
+  int _amount;
   int _paymentType;
 
   PaymentModel(this._icon, this._color, this._name, this._date, this._hour,
@@ -16,7 +16,7 @@ class PaymentModel {
 
   String get hour => _hour;
 
-  double get amount => _amount;
+  int get amount => _amount;
 
   int get type => _paymentType;
